+++ CHANGES.rst
@@ -5,9 +5,11 @@ Version 5.0.0
 -------------
 
 * Adding support for msgpack
+* Adding `default_box_no_key_error` option to not raise key errors on `pop` `popitems` or `del` (thanks to Patrock)
 * Adding `mypy` type checking, `black` formatting and other checks on commit
 * Change back to no required external imports
 * Removing readthedocs generation
+* Removing overrides for `keys`, `values` and `items` which will return views again
 
 Version 4.2.2
 -------------